fixing

+++ setup_update.sh
@@ -1,8 +1,16 @@
 #!/bin/bash
 export LANG=""
+echo "Checking internet connection"
 echo "-----------------------------------------------"
-echo "Updating scripts - Please wait..."
 
+host1=github.com
+if [ ping -w5 -c3 $host1 > /dev/null 2>&1 ]; then
+	echo "Updating scripts - Please wait..."
 	git reset --hard HEAD
-	    git pull https://WilliamStam:awssmudge1@github.com/WilliamStam/Impreshin-Setup.git master
+    git pull https://WilliamStam:awssmudge1@github.com/WilliamStam/Impreshin-Setup.git master
+else
+	echo "no connection"
+	sleep 1
+fi
+
 bash ./setup.sh
